+++ crates/nu-cmd-dataframe/src/dataframe/eager/to_nu.rs
@@ -5,6 +5,8 @@ use nu_protocol::{
     Category, Example, PipelineData, ShellError, Signature, Span, SyntaxShape, Type, Value,
 };
 
+use crate::dataframe::values::NuExpression;
+
 use super::super::values::NuDataFrame;
 
 #[derive(Clone)]
@@ -16,7 +18,7 @@ impl Command for ToNu {
     }
 
     fn usage(&self) -> &str {
-        "Converts a section of the dataframe into nushell Table."
+        "Converts a dataframe or an expression into into nushell value for access ane exploration."
     }
 
     fn signature(&self) -> Signature {
@@ -28,7 +30,11 @@ impl Command for ToNu {
                 Some('n'),
             )
             .switch("tail", "shows tail rows", Some('t'))
-            .input_output_type(Type::Custom("dataframe".into()), Type::Any)
+            .input_output_types(vec![
+                (Type::Custom("expression".into()), Type::Any),
+                (Type::Custom("dataframe".into()), Type::Table(vec![])),
+            ])
+            //.input_output_type(Type::Any, Type::Any)
             .category(Category::Custom("dataframe".into()))
     }
 
@@ -67,6 +73,15 @@ impl Command for ToNu {
                     span: Span::test_data(),
                 }),
             },
+            Example {
+                description: "Convert a col expression into a nushell value",
+                example: "dfr col a | dfr into-nu",
+                result: Some(Value::Record {
+                    cols: vec!["expr".into(), "value".into()],
+                    vals: vec![Value::test_string("column"), Value::test_string("a")],
+                    span: Span::test_data(),
+                }),
+            },
         ]
     }
 
@@ -77,20 +92,25 @@ impl Command for ToNu {
         call: &Call,
         input: PipelineData,
     ) -> Result<PipelineData, ShellError> {
-        command(engine_state, stack, call, input)
+        let value = input.into_value(call.head);
+        if NuDataFrame::can_downcast(&value) {
+            dataframe_command(engine_state, stack, call, value)
+        } else {
+            expression_command(call, value)
+        }
     }
 }
 
-fn command(
+fn dataframe_command(
     engine_state: &EngineState,
     stack: &mut Stack,
     call: &Call,
-    input: PipelineData,
+    input: Value,
 ) -> Result<PipelineData, ShellError> {
     let rows: Option<usize> = call.get_flag(engine_state, stack, "rows")?;
     let tail: bool = call.has_flag("tail");
 
-    let df = NuDataFrame::try_from_pipeline(input, call.head)?;
+    let df = NuDataFrame::try_from_value(input)?;
 
     let values = if tail {
         df.tail(rows, call.head)?
@@ -110,14 +130,26 @@ fn command(
 
     Ok(PipelineData::Value(value, None))
 }
+fn expression_command(call: &Call, input: Value) -> Result<PipelineData, ShellError> {
+    let expr = NuExpression::try_from_value(input)?;
+    let value = expr.to_value(call.head);
+
+    Ok(PipelineData::Value(value, None))
+}
 
 #[cfg(test)]
 mod test {
+    use super::super::super::expressions::ExprCol;
     use super::super::super::test_dataframe::test_dataframe;
     use super::*;
 
     #[test]
-    fn test_examples() {
+    fn test_examples_dataframe_input() {
         test_dataframe(vec![Box::new(ToNu {})])
+    }
+
+    #[test]
+    fn test_examples_expression_input() {
+        test_dataframe(vec![Box::new(ToNu {}), Box::new(ExprCol {})])
     }
 }
