+++ crates/nu-cmd-dataframe/src/dataframe/expressions/mod.rs
@@ -1,6 +1,5 @@
 mod alias;
 mod arg_where;
-mod as_nu;
 mod col;
 mod concat_str;
 mod datepart;
@@ -15,7 +14,6 @@ use nu_protocol::engine::StateWorkingSet;
 
 pub(crate) use crate::dataframe::expressions::alias::ExprAlias;
 use crate::dataframe::expressions::arg_where::ExprArgWhere;
-use crate::dataframe::expressions::as_nu::ExprAsNu;
 pub(super) use crate::dataframe::expressions::col::ExprCol;
 pub(super) use crate::dataframe::expressions::concat_str::ExprConcatStr;
 pub(crate) use crate::dataframe::expressions::datepart::ExprDatePart;
@@ -44,7 +42,6 @@ pub fn add_expressions(working_set: &mut StateWorkingSet) {
         ExprConcatStr,
         ExprCount,
         ExprLit,
-        ExprAsNu,
         ExprWhen,
         ExprOtherwise,
         ExprQuantile,
